Fix some issues from refactor
